"3:10pm. https://stackoverflow.com/questions/16959539/differences-between-self-and-smalltalk
This is actually pretty interesting. Maybe I could base Self as a template for OO in Spiral.

I mean, Spiral already as it is can do OO, but it is a mystery how a language like it would provide a first rate IDE experience.

3:20pm. https://youtu.be/3ka4KY7TMTU?t=2451

"
Compilers are for the 1950s and 1960s.
* A compromise for machines with small memory and magnetic tape.
In service of creativity, implementation must create illusion of reality
* Principle: offload cognitive burden
* Practice: 1/3-second turnaround, source level-debugging, programmer can factor without worry.
"

3:55pm. https://youtu.be/3ka4KY7TMTU?t=4137

"(on artists vs entertainers) 'Your ultimate goal might even be religious, not simple pleasure, but transcendence.

Now your weak point is self indulgence. That is a tendency to think that every lame or trivial idea that comes into your head is important just because the muse brought it along with the good ones. Remember that an artist needs an audience too and a ruthless internal editor to make sure you work is as good and focused and as accessible that can be on its own terms.' And Joe goes on to write: 'Even as an entertainer, I had a knack of making things difficult for myself by trying desperately to do them as brilliantly as possible. Maybe it was the artist in me.'"

4pm. Let me finally resume the tutorial. At this rate it seems I might only get 2 chapter out of the way. How could this happen?

Well, this video helped me come to a conclusion on where exactly I should go with Spiral. Before I do anything else, the way to make it more tool friendly, what I should do is try and visualize it. Visualize all the aspects of the compilation process. Rather than making it more OO or something else, that is the general trajectory I should pursue.

Do not target the user experience the directly. Target the compiler itself instead.

4:10pm. Focus me. Stop with the social media. I got up so early today, just where is the time going to? Focus!

4:15pm.

```
 | dot dotPosition |
 dot := $..
 dotPosition := (self size to: 1 by: -1) detect: [ :i | (self at: i) = dot ] ifNone: [^ ''].
 ^ self copyFrom: dotPosition + 1 to: self size
```

Actually, this is one of the things I've been wondering about. How would the return statements even work from within blocks?

...Ah, this works? Really? How nice.

This is actually pretty cool in my book. Ok, I've changed my mind on explicit returns. But it would be a bit hard to do in Spiral as I'd need the ability to clean up the stack. And it would not work across language boundaries. So I'll leave it off the wish list.

4:30pm.

```
Pharo is a multi-threaded system, and there are many lightweight processes (also known as threads) running concurrently in your image. In the future the Pharo virtual machine may take advantage of multiple processors when they are available, but at present, concurrency is implemented by timeslicing.
```

Although it is a pity it does not have this sort of concurrency, it is more important that the GUI goes not hang the interpreter than anything else.

Let me play with the process explorer. I am really liking the whole Smalltalk system. This is what I want to become; I want to internalize Pharo - all of it.

4:40pm. 161/376. Finally the subject comes to source control. Yesterday I got back mail from Ducasse telling me clear out my cookies and ask on Discord. But rather than interfacing with that Beta tier site, I think I'll make Github work for me.

"Up to Pharo 5.0, this is done using a tool called Monticello, and is a much more powerful and effective way to work, especially when working in a team. In the future Pharo will officially offer support for Git and hosting on servers such as GitHub, BitBucket and more."

Ah, I see, so that is on the roadmap.

4:50pm. 162/376. This stuff is interesting. So it is possible to extend a class using methods defined in some other package?

5:10pm. Focus me. I am surfing /a/ again.

6pm. 178/376. I can't believe it is this late. I only covered one chapter today, but it was a big one. It seems I am halfway into the book now.

I do not feel like it anymore, so let me stop here. Lunch should be soon anyway."