"10:05am. I had trouble sleeping tonight. I was thinking the whole time about how I would do OO properly in Spiral. At first my thoughts were just around replacing type literals with keyword messages - I came to the conclusion that this is a good idea and worth doing. It would significantly improve the readability of the language. But after that my thought turned to how I could do objects directly. I can do them somewhat now using functions and modules, but if I did them in the language I could compile them much more effectively. What I would do is combine them with keyword messages and make their evaluation lazy. That would allow me to replace `inl rec` with them as well.
Once I have such inbuilt objects and keywords directly in the language, I could make Spiral an OO language. `.unary_message` instead of type literals and `keyword1: a keyword2: b` for such kinds of messages. I could easily alias them to integers during parsing and get significantly faster lookups.

Spiral being a true OO language means that the primitives would be objects. That would be cool. But a that point, even without inbuilt object this undertaking would turn from a speedup to a complete language redesign.

10:20am. Out of everything that I've seen about Smalltalk, keyword messages are such a stroke of brilliance. Right now I am definitely of the opinion that using modules in place of them is a design mistake in the language. The reason for that being is that I am often doing things like...

```
match x with
| {state} -> ...
| _ -> ()
```

It is easy to make a mistake when doing this and write `{staet=...}`. Messages on the other hand match exactly.

Once I do this, the main difference between modules and objects is that modules will be strictly evaluated and allow only unary messages as fields.

10:30am. This should streamline the language and give that last spurt of speed that I need.

It would be a massive undertaking as I'd need to rewrite not only the language itself, but the libraries as well. So about 2 months of full time work. And I haven't decided on all the details of how I want to do it. Object updates still feel a bit tricky to do effectively.

Also there is Pharo looming in the background. To really get all that I want I should rewrite Spiral in Pharo and take advantage of its interactivity to implement a type debugger.

10:35am. Ah, what a pain. Will I ever be finished with this? What am I even fighting for here? What sane reason is there to go so far just so I could get good GUI experience?

All the ML libraries are in Python, but here I would be starting from scratch once more.

It is true that Pharo is a far superior language to Python - it is because of that that my heart is pulling me in that direction.

10:40am. Programming is my art. I do not know where it happened. I tried not to let it happen by focusing on a real world task, but at some point laying out the best foundations possible took over from all the other concerns. Rather than winning the battle, all I care now is doing proper exercise.

It is not going to work...for them. Nobody is going to cause the Singularity from the command line. Even if both the hardware and the algorithm were here, it is not like it would be possible to just press 'Run' and watch the AI swallow the universe. Rather, at that point the game really starts. What is going on now is just the prelude to it.

10:45am. So I might as well enjoy it. The desperate feeling of trying to get better at programming and the search for every slight edge to add to my arsenal.

10:50am. Reality is not so kind that it will just dump what I need right into my lap. I do not deserve it at the moment.

I will earn it by completing Spiral. I will drain out all of its slowness and bring outs it full potential. I will boost its productivity to the limit with an IDE and to top it off, merge it with a dynamic language. At that point the convergence will be complete and Spiral will have evolved to the limit of what a human can do.

I'll have my ML battle station at that point.

Rather than thinking of how much work this would be for an indeterminate gain, would it not be more fitting for me to focus on how great it will be when all the pieces come together? AI is something that I should always have been doing, and at the end of this deep learning wave a new way of programming will be born. I am a fool. I should have foreseen the future and set out on this path at the beginning of high school. But I guess I really did need to take a decade to solve the purpose of life and understand the truth of the self improvement loop.

11:05am. Since I did not get much sleep and it is this late, let me have breakfast here. I will resume my studies after that.

I am really looking forward to having keyword messages and proper objects in Spiral, but I will leave work for that after I am done with Pharo which will take a while. I have a whole bunch of things I want to learn before I do anything concrete with the language, though as always I will be thinking about it.

In the last eight months, after working on Spiral for a long time I finally got a real taste of RL for once, though it is far too little to satisfy me. I'll have a proper mean when I am done with all of this."