"2:15pm. I am thinking of how I could do inheritance in Spiral. Even with prototypal OO, I'd need some way of calling `super`.
Well, is it that hard?

```
inl SomeClass(super, values) =
    [super with
    methods ...
    ]
```

The issue with this is that if I user `super` directly, then `self` won't get interpreted as it should.

2:25pm. ...So for super I should evaluate super from the parent class, but pass the self from the leaf class?

That is actually quite brilliant. It could be done.

Actually, this is a much better solution than the one that I came to over the night.

If I put in explicit inheritance rather than copying everything, that will also speed things up. In the book just now it mentions that Object in Pharo has 400 methods. Do not want to copy all of that on every `[super with ...]`.

Also, for once rather than `[super with ...]` I shold also allow traits. So...

```
super
    uses: traits
    with: [
        methods...
    ]
```

This is how it should be.

That having said, how exactly am I going to use this sort of thing? Should the prototype object have `uses:with:` and `with:` inbuilt? Seems like the most sensible thing.

2:35pm. No the `[...]` should not have anything built in by default. Rather in `Core` I will define `ProtoObject` and `Object` with the two inheritance methods.

2:45pm. Yeah, that would work. Ok, I think I understand what was meant in the book when it said that `super` is more static than `self`. I can visualize the search now.

2:50pm. OO would really be super useful for Spiral as I do not have any real way of passing typeclasses directly and such.

But forget this sort of thing for now. Before I decide upon anything I need to get a lot more experience with a real OO language like Pharo.

2:55pm. One thing that also occurs to me is that if I want to write things like

`(1 to: 3)` then the `Integer` class would need a way to reaching `Duration` which would in all likelihood had to be defined after it. So I have no easy way to express this in Spiral.

The best thing would be to go all the way into OO in that case. That would have the added benefit of exposing a lot of structure to the tools.

3:20pm. Let me stop here. I need to do the chores.

3:50pm. Phew, let me resume.

4:10pm. 208/376. This chapter seems like it will be a long one, but it is just a summary of the standard library so I do not have to dwell too much on it.

I am not sure what I should do...

I understand how to deal with single inheritance with objects completely now. I've worked out `super`. Traits won't be any trouble.

But to make a true OO language, everything needs to be a class hierarchy. Spiral on the other hand is one long flow.

4:15pm. Even more than that, I am not sure whether I even want OO instead of the long flow.

That is one of the issues that is troubling me right now. I just do not have enough understanding to know what I want.

4:45pm. Just how is `1 to: 3` implemented. I can't find the method `to:` anywhere.

Ah, it is in Number.

```
to: stop
 "Answer an Interval from the receiver up to the argument, stop,
 incrementing by 1."

 ^Interval from: self to: stop by: 1
```

Yeah, this is exactly how I thought it would be.

5:40pm. 237/376. Let me call it a day here. I had enough.

5:45pm. Ough how annoying all this is. I mean figuring out how to do OO for Spiral, not so much studying Pharo.

I can make inheritance work, I can make `super` work, but I can't really nest classes...

The reason why Pharo needs things like metaclasses is because it cannot do everything on the top.

Spiral does not need them, but on the other hand it cannot have Integer call `Interval` and such...

Agh, I cannot deal with this. I am really calling it a day here."