"10:05am. https://arxiv.org/abs/1805.08195 Depth-Limited Solving for Imperfect-Information Games
I woke up much earlier, but I did not feel like getting out of bed.

"A fundamental challenge in imperfect-information games is that states do not have well-defined values. As a result, depth-limited search algorithms used in single-agent settings and perfect-information games do not apply. This paper introduces a principled way to conduct depth-limited solving in imperfect-information games by allowing the opponent to choose among a number of strategies for the remainder of the game at the depth limit. Each one of these strategies results in a different set of values for leaf nodes. This forces an agent to be robust to the different strategies an opponent may employ. We demonstrate the effectiveness of this approach by building a master-level heads-up no-limit Texas hold'em poker AI that defeats two prior top agents using only a 4-core CPU and 16 GB of memory. Developing such a powerful agent would have previously required a supercomputer."

Maybe I should just consider implementing this? I've not even looked at its contents.

10:15am. As I study Pharo I have a great inner struggle with regards to what the optimal style for Spiral is.

Right now it feels like the standard library is a mut of different styles.

10:20am. Functional on top of OO? That is the thing that is popping into my head currently. I do not want to part with elegance of FP - I want to be able to write `fold (+) 0 x`, but I also want the consistency of OO. With regards to the `Interval` thing I can just have the module extend the primitive types.

10:30am. I am still thinking about it. Having `+` be `inl (+) a b = a add: b` does not sound that bad.

Having functions be these kinds of wrappers is not a bad idea. It will give me the kind of polymorphism I could only otherwise get with typeclasses.

Note that lack of expressiveness was a problem in place. For example, because of that I had to implement activation functions in `CudaAD` module.

10:40am. Yeah, I can see it. Adding objects and then making functions kind of a wrapper around them would be quite an interesting way of going about it.

10:50am. It would also make things that much slower. Also, I am still not sure whether I want inheritance at all. But the whole message passing thing did capture my imagination.

10:55am. Before I decide upon anything there is something I want to make sure of. Is there really a use case for a superclass calling a subclass? This is the default behavior with inheritance that could occur with some super calls. Or will my use case really always be covered with composition?

I'll admit I see plenty of reason why a message would be passed upwards in the chain, but not much reason to see that happen in the opposite direction.

11am. I really should be reading the Pharo book, but I am too distracted with my own thoughts.

Let me take a break here. It is time for Tog."