"12:20pm. 243/376. I'll admit I won't really focus on this too much. I just want to get through this phase and gather my determination.
12:25pm. 251/376. Here is a chapter on Morphic. Now this is why I picked up Pharo in the first place but..

I feel a deep insecurity right now.

Just what are my principles? Why did it come to this?

I wanted to do it all using Spiral, right?

Just why is my reward for doing all this work even more work? Will it ever end?

If I were to describe what I feel, and what I've been forced to endure for the past four years, it would be humiliation and resentment.

It is just so hard, doing what I am doing without any help. I have to keep a far reaching global view, am not allowed in any way to compromise and yet do the lowliest kind of slave work and put every piece into place by my own hands. I am both the highest of the high and the lowest of the low.

Damn that is hard.

Trading I failed because I compromised too much.

I should have gone for it with machines, but back then I never wanted to program ever again. I wanted to do it with my own power.

12:40pm. I should have known that I would never be able to avoid the quest to grasp the root of everything.

Desire. Just what is desire? How far can it go? What do I need to do to grasp its secrets.

All the temptations have driven me into perplexity.

I might have succeeded in attain a measure of skill in staged functional programming, but that by itself has not given me the power that I ultimately sought. It is pretty much the same as trading.

In trading I found so many different perspectives and learned so many things, but in the end they were all useless. How is it possible that all those patterns and all that stupidity is something I cannot take advantage of directly? Is selling garbage to idiots the only way to make a decent living as a human?

And is attaining the power of machines the one true way?

12:50pm. Let me just let it all out. I am so sick of this.

Yeah, I made a language, I got pretty good at the kind of programming nobody else at this time understands, but the power eludes me.

All the real power is held by the connector languages. It is kind of how Twitter and Facebook and Youtube hold disproportionate sway over the content creators. It seems to be like a principle of reality brought to life. Call it the principle of hierarchy.

Ah, how I hate it.

And the thing I hate the most about it is that I do not understand it. If I did, I could formalize it into code and make its power my own.

12:55pm. I am just so fucking pissed that it has come to this.

Yeah, I know what I need to do at the rational level - master a fully integrated dynamic language like Pharo, connect it to Spiral, make the ML library. It does not matter what has happened before, what matters is that I do it right. And just like I studied Haskell and other typed FP languages for a bit while having F# as my main focus back in 2016, I should round that off by studying Python, Perl and Ruby for kicks.

It does not matter if it takes 2-3 years. What matters is that I do it right.

It does not matter if I go for another 2-3 years before making a dime, what matters is that I keep increasing my mastery of programming.

Just like in trading I deep down understood that there were a dozen reasons why I might not succeed despite all the desire, so it is with Spiral that I understood that statically typed languages cannot really compete with the sheer expressiveness of dynamically typed ones. It is even worse than that, given that MLs like F# and Haskell are dynamically typed languages under the hood, there is no real reason to think they would be slower either.

I just do not understand - where does that put type systems? As niche tools?

Agggghhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhh...

I hate and hate and I hate...is this world really going to drive me this far? When the 2022 deadline comes around will I still be working on languages?

Even as I rage inside, I can feel a sliver of desire to do that work and to attain the full mastery getting ready to sprout.

1:15pm. I am just such a fool. I understand nothing. Am I even better that the average Javascript programmer at this very moment?

I need more time. I need three more years. I am not even close to being done. I am really sorry about that. The fact that attained no real power is unforgivable, yet I must go on.

I know that time is expensive and that as a human I can just keep taking these kinds of loans indefinitely, but I must ask for more. I know that the price for them will be my life, but I must do it for a little while longer.

Even if I put my all into it, just what could I achieve in the next month? Nothing. It will have to be years.

It is so embarrassing. My desire is just so weak, if it was stronger it would not have had turned out like this. I need to make my will felt.

...

I hate it so much that I might explode, but I believe in my plan. That is why I suffer so much now. It seems switching gears is always accompanied by that.

1:25pm. It cannot be helped.

For the next few months, let me just focus on dynamic languages while I think of what I want to do with Spiral on the side and maybe do some light work on it, just enough to allow myself to keep posting on the PL sub.

That way if not just my monitor, but my entire rig goes out I'll be ready for the worst case which is having to do paid work. I've done all I can to avoid falling into that as it would kill the quest, but bad luck can always come around. If I was the chosen one I would not be in this situation right now.

If I really was the player, there is no way I could have possibly had such an irrational trajectory through life. Somebody real would have more control over his desires.

1:35pm. Let me resume the Pharo By Example book. Sure, it is not quite the power I wanted, but dynamic languages are definitely the real deal.

Compared to making Spiral, the degree of power mastering one would receive especially given how little it would take is truly great.

I am by nature that kind of person to go after higher hanging fruit first, but Pharo, Python, maybe add some web language to the mix would truly be an indulgence.

1:45pm. Let me go on a journey. GUIs, games, webdev, FFIs, JITs in dynamic languages...there are all sorts of things I did not cover in 2015-2016 during my study sessions. I did make a (private) poker game in F#, but the GUI programming there was just so nasty. It turned a 300 line program into a 1500 line one. F# really is not suitable for this kind of work and for GUIs to be ergonomic, much better will be needed.

GUIs are rather important to me - before I began work on the game, I spent almost a month going through Petzold's book and trying out trivial examples by hand. I am going to pit Pharo against Python, Ruby and the rest in both GUIs and webdev. I will make the language of my choice the one which gets the highest scores in terms of interactive experience.

I am assuming that will be Pharo, but I should do it properly.

3 months...I will give this 3 months. It will probably take less, but I'll allot myself 3. That will give me ample time to get familiar with all the subjects of interest. Then I will resume work on Spiral. With a better understanding of what exactly is it that I want, I will redesign it with keywords arguments, objects and whatever with the intention of making it an embedded language for whatever is the dynamic language of my choice.

I will make it persistent and make that type debugger for it.

I will do whatever is needed until I stop feeling shame.

I will make it properly interactive much like the ML library I am going to make.

Whatever I do, the overarching theme for the next 3 years will be to break the dependency on the command line.

I swear I will make that happen."