The journey takes an unexpected twist.
https://youtu.be/baxtyeFVn3w
Lambda World 2018 - What FP can learn from Smalltalk by Aditya Siram

Yesterday I watched this talk and it cheered me up. Most of the stuff he showed me was actually new to me. Library specific visualizers and debuggers? That is exactly what I want! And Petitparser which I did not know existed was a parser combinator library with such functionality. I was wondering if Pharo even had that and it turns out that the situation was much better than I imagined.

Pharo it is. I have no real alternative to Smalltalk. I cannot compromise here if I want to do a professional job.

I do not know how long it will take me to do a ML library the way I really want, but right is right and might is might. They both point in a clear direction and do not care how much effort it would be to do this.

I am much further from my goal than I ever imagined, but at the very least I have a chance of doing it properly. Programming to me is definitely not about bragging. Even if Spiral's design is revolutionary, I cannot say that a language that is so tedious to program in that I seriously considered dumping it in favor of Python is something I can accept.

Yeah, I can definitely do a generic map kernel properly now. And I have the ML library foundations set.

Though this is not the kind of work I expected I would be doing I will have to as there is nobody else. I looked down on it and hoped that somebody else would pick up the slack for doing this, but it seems nobody in the whole world seems to have the unique intersection of interests that I happen to have at the moment. There is nobody else who can do this - unless I take up the role, in the 20s the Singularity will literally be tried from the command line.

More than that - forget the Singularity. Do I or anyone really wants to use programming languages like Spiral is now - from the command line? Of course not.

Moreso than Spiral, unless I set out on this path, I am going to fall into a local minima where all of my programs are like this.

I need a language that can do user interaction seamlessly.

And more than that, I having made Spiral, I am truly amazed how fast Pharo is. Just what is needed for a language to be so fast? What am I missing in Spiral's implementation? Even if I did the latest changes, I do not think I could make it that fast during compilation.

I want to understand Pharo in depth.

Though I'd rather get 'practical', ignoring sheer raw ability, I have plenty of vectors left open in which I can evolve, none of which I will be able to tap into if I stick with how I've been doing it up to now. I need to change.

At least what I will be doing from here will be more fun that what I've done previously. I had enough of the kind of programming that I did. I need a breath of fresh air.