"4:10pm. https://youtu.be/lRQ447XI6Hc?list=PL4actYd6bfnw7ous5We-XnWq0NZquP3gM&t=698
I still have no idea what `yourself` should be in this context.

Forget it, let me go back to the book.

The Seaside demo is quite interesting. I could build a webpage in this.

4:35pm. Ah, I get it. I'll remove automatic unboxing and use the # operator to trigger it instead. That should do the trick. Actually it is pretty amazing that this never occurred to me.

4:40pm. 306/376. I think I will just skip this particular exercise as it seems like it could take a while.

4:45pm. 313/376. Let me stop here for a bit as I need to take a break. Actually, maybe I will call it a day here.

Only two chapters left in the By Example book. I admit that I pretty much skipped the second half of Seaside.

After I am done with the book, I'll follow up into the Deep Dive one.

5pm. https://github.com/pharo-open-documentation/pharo-wiki/blob/master/General/Traits.md

'Since Pharo 7 Traits are modular and not tight the the Kernel. So it would be possible to have multiple implementations.'

Note the 'the the'. I already found a grammar error in the changelog.

`if the do not override it.` should be `if they do not override it`.

`Traits are composabls` should be `Traits are composable`

5:05pm. `Pharo is now provided in 64-bit version in Linux and macOS and brings even better performance and stability (beware, 64bits version is a new technology and a small amount of tests is still failing)`

Should be `are still failing`.

5:30pm. Done with lunch. Let me make that fix.

5:45pm. Found another one by accident, but it is for the 2.0 version and is deeply buried so I think I'll just spare the authors the effort.

I had enough of this for today."