"9:15am. I am up. Today snow feel outside and I really feel rested. I think I am going to be able to focus on Pharo today as I am no longer obsessed on how to fit objects into Spiral. I think I have a good grasp on it now.
9:20am. https://news.ycombinator.com/item?id=18969155

">I couldn't understand the fascination with the integrated environments though.
The only thing I can recommend it watching some Alan Kay videos or Pharo demos.

Windows and Linux UI conditioned everyone to think of UI as a crutch for incompetent users, rather than a tool useful in its own right. This cannot be undone in a single reply, no matter how elaborate it will be.

Traditional tools are definitely more polished at the moment, but they are exactly that - traditional, shaped by historic accidents, rather than design. Software development is currently running on a set of highly refined steam engines.

For example, what are the chances that a simulation of a teletype is the best interface for interacting with an operating system? What are the chances that the best conceptual model for a unit of information is manila folder or a phone book ("files", "directories").

Also, IDE came to mean "bloated monolith for editing and compiling code". But it should really mean that application interface is the development tool. Tools that are integrated into OS and applications."

Right now I am reading this thread. Let me chill a bit and then I will start with the book.

I will be able to focus today.

I'll finally finish the book and get started on the deep dive.

9:40am. Let me finally start. I can always slack later.

10:15am. I deduced during the night that `yourself` would just be `^self` and I see that I was right.

10:45am. "[:aClass |
 SystemNavigation default
  browseMessageList: (aClass withAllSubclasses gather: [ :each |
   each methodDict associations
    select: [ :assoc | assoc value sendsToSuper  ]
    thenCollect: [:assoc | RGMethodDefinition realClass: each selector: assoc key] ] )
   name: 'Supersends of ', aClass name, ' and its subclasses'
 ]
 value: Collection"

Keyword arguments really do make all of this quite readable. I should enshrine this lesson in the next version of Spiral.

11:50am. 331/376. Let me stop here as I need to eat. The stuff in this chapter is quite interesting, especially the chapter on generating code that I am about to read.

I am definitely making good progress through the book, though I am started to get distracted by thinking of how to implement objects (again). It does not matter, I really should defer this kind of thinking when the implementation time comes around. I already went through all of this in the past week.

Let me rest for a bit."