"12:40pm. https://youtu.be/dTPI6wh-Lr0?t=1566 "If we looked at what's the mapping in between what we think the reality is in here and what we think it is out there, we'd see a complete mismatch. I recall that being definition of not being sane. Sanity without normalization should be something like how good is the mapping. But instead we do not define it that way. We define insanity or mental illness simply outside what most people actually do no matter how crazy it is."
I did not think that this talk by Kay would be on the issue of mental health.

"Adults are dangerous, really, really dangerous."

Yeah, this is the thought I had myself as a kid, though I was smart enough not to need to grow up during a war to see it. The decision to make that thought a core part of my identity explains a lot of the way my personality has developed.

If I had never thought had this thought I would have probably been able to fit into society, instead of spending all my energy into trying to engineer around it.

Mistrust towards society due to its lack of self awareness and the rational conclusion that omnipotence itself is the true purpose of life are the definition of me and I have steadfastly resisted all efforts to be defined otherwise.

Well, that is how it was, but the last ten years have changed me. Compared to how I was before, I now have a deep loathing for myself for not being able to win at trading and for not being able to defeat those I saw as weak. And realizing the truth of the self improvement loop gave me an edge. It sharpened what was always lurking deep within me.

I now trust in my shadow. My hatred, despair and insanity will bring the Spiral project to its conclusion. I do not care if I have to blast open the rift between the worlds to do it - I will make the power of the machines my own. The time spent will be my apology for always believing, but never doing anything with them. I will be the one to resolve all the contradictions and I will be the one to grasp what lies at the end of the journey.

Let me go forward with the talk. I will finish it and then finish the first Pharo book.

1:15pm. https://youtu.be/dTPI6wh-Lr0?t=2684

I wish I'd learned programming like this.

1:35pm. I should have skipped the second half. This sort of thing always turns into liberal hogwash. The talk is for that sort of audience and Alan Kay is that sort of type as well. Well, it is classical feel good hoghwash he grew up with rather than the recent adversarial racial identity motivated liberalism.

1:45pm. I think possibly the biggest and most adverse delusion currently embedded in Western collective consciousness is that biases can be combated using conditioning. I fell for that one hard myself as a matter of fact, but the record of the past 20 years of my life speaks for itself. It turns out to be nothing more than window dressing. It evaporates as soon as willpower gets exhausted.

Rather trying to get better using those routes, I think that better tools are a more effective way to go about it. And since tools are the solution to every problem the only real skill worth having is tool invention.

Had it not been for the whole Singularity thing looming in the background, I think that organized education would have been devoured by the Internet in due time. Though that is too optimistic given that schools also double as kindergardens, so me saying this is just my bias given voice. Well as an alternative, society as a whole getting devoured is on the table.

1:50pm. Let me finally resume the book...er, after I finish the chapter.

2pm. "The other most common application of intercepting not understood messages is to dynamically load or generate the missing methods. Consider a
very large library of classes with many methods. Instead of loading the entire library, we could load a stub for each class in the library. "

I did something in the old F# library using laziness.

2:05pm. 331/376. Hmm, actually are messages also objects?

2:45pm. "Method wrappers can be used to perform any kind of suitable behaviour before or after the normal operation of a method. Typical applications are instrumentation (collecting statistics about the calling patterns of methods), checking optional pre- and post-conditions, and memoization (optionally cacheing computed values of methods)."

I am guessing this sort of thing would be good for profiling something like a compiler. This was one of the hard things in Spiral.

2:50pm. 339/376. Only a little bit more is left.

Let me stop for a bit here though as I need to do the chores as is usual at this time.

3:35pm. Phew, time to resume. The next chapter is on regular expressions.

3:50pm. I am trying to remove `Webdir` or rename it and am getting the error message that `ObjectsAsMethods` does not understand something. Should I just get a new image? I am not sure what I should do here. Maybe I should move to 7.0?

Let me do it.

3:55pm. It seems that going for the bleeding edge build is not good, but instead I should just get the 7.0 development version. It has been a few years since 6.0 so I might as well. At the very least 7.0's Iceberg does not give memory errors. I really should go for 7.0 since I'll want to use Git rather than the unstable Smalltalkhub. I did read that Git will be the primary versioning system for Pharo going forward.

Ok...Let me see if I can get my focus on. I want to finish the Pharo By Example book today so I can start the Deep Dive one. I'll think of what I want to do after that.

4:40pm. 347/376. Ok, let me just start skimming the Regex stuff. I do not have a website that I could test this stuff on. I am not that interested in Regexes anyway apart from how they are compiled. They are the ideal target for partial evaluation, but I doubt I will ever be using them in my life.

4:50pm. "The matcher is similar in spirit, but not in design to Henry Spencer’s original regular expression implementation in C. The focus is on simplicity, not on efficiency."

Maybe once Spiral for Pharo is done, I might be able to do something about this.

363/376. I am nearly done with the book and for the day as well. The Regex chapter was really long it seems.

5:05pm. 366/376. What is this `ObjectFinalizerCollection` here.

5:20pm. I am done with Pharo By Example. I did not feel like the MOOC again so I did the book.

5:25pm. Let me take a look at Deep Dive which is an expert book and I will call it a day here."