One thought that keeps coming to me is that making a ML library for Pharo would be a lot easier than giving Python the deep integration that Pharo offers.
Pharo is unique in the sea of indistinct scripting languages. It is worth pushing forward with this plan.

Python makes sense in the short term as it has all the best libraries, but going with it in the long term would be a disaster once real AI agents arrive on the scene.

I am going to need Pharo's live coding functionality to handle the coming challenges.

The plan is thus - since I do not feel like doing anything else in Pharo, the thing that would make the most sense for the sake of attaining familiarity with the language and for the future steps would be to start work on translating Spiral into Pharo immediately.

That would be a real, rough challenge.

Although by translate Spiral, I really mean the unfinished v0.1 of it. To be honest, I'd rather not have to learn Pharo and do the translation on the unfinished version all at the same time, so I am going to have to do at least v0.1 in F# as per plan.

The compiler is a difficult sort of thing where TDD is not exactly going to help. I am going to have to do like 4k lines without doing a single test...

Yeah, just like when I started Spiral, I think it might be better to do a hugely pared down version at first.

The most important thing is that I steel my will for what is going to come.

The next two years I will spend in order to perfect Spiral and make Pharo an AI suitable language will be the beginning of the end.

Today there is the SC2 demo by Deepmind. I am quite skeptical that they managed to crack unsupervised learning so soon - I'd expect it to happen in the SL domain first, but at some point it will happen.

I still have time to do things right, so I should take advantage of it.

Damn, I can't believe I will be working on Spiral for this long. It is amazing, but a lot of the widely used languages have been in development for over a decade before they achieved widespear adoption.